backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.9.0
APScheduler==3.11.0
asttokens==3.0.0
async-timeout==5.0.1
asyncua==1.1.5
attrs==25.3.0
av==12.3.0
backoff==2.2.1
boto3==1.35.60
botocore==1.35.99
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
deepgram-sdk==3.10.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.110.3
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
google-crc32c==1.7.0
groq==0.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.7
ifaddr==0.2.0
imageio==2.37.0
inference==0.42.1
iniconfig==2.1.0
ipykernel==6.29.4
ipython==8.34.0
jedi==0.19.2
jiter==0.9.0
jmespath==1.0.1
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
marshmallow==3.26.1
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==12.570.86
onnxruntime==1.19.2
openai==1.67.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
pillow_heif==0.22.0
platformdirs==4.3.7
pluggy==1.5.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybase64==1.0.2
pycparser==2.22
pydantic==2.7.4
pydantic-core==2.18.4
pydantic-settings==2.8.1
pydot==2.0.0
pyaudio==0.2.14
pydub
pyee==13.0.0
pylogix==1.0.5
PyJWT==2.10.1
pymodbus==3.8.3
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyreadline3==3.5.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==310
pyzmq==26.3.0
redis==5.0.8
requests==2.32.2
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.0.1
roboflow==1.1.58
s3transfer==0.10.4
scikit-image==0.24.0
scipy==1.15.2
seaborn==0.13.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.37.2
structlog==24.4.0
supervision==0.25.1
sympy==1.13.3
tifffile==2025.3.13
tldextract==5.1.3
tokenizers==0.21.1
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.3.8
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
ultralytics==8.3.88
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn[standard]==0.29.0
wcwidth==0.2.13
websockets>=12.0
yarl==1.18.3
zxing-cpp==2.2.0

# LangChain + Pinecone + PDF/CSV additions
langchain==0.3.23
langchain-community>=0.0.29
langchain-core==0.3.51
pinecone-client>=3.0.0
cohere>=4.0.0
tiktoken
pypdf
langchain-cohere>=0.0.2
langchain-pinecone>=0.0.3
pyaudio 
opencv-python
mediapipe
scipy
numpy
protobuf==3.20.3
transformers
sentencepiece
